Thu, Feb 12 2004:
Updated Evaluation to include standard deviation of the estimated accuracy, per class.  Note that this can only work in the overall evaluation panel.  At the moment I was not able to differentiate the overall panel's summmaryStatistics (String) vs. the single fold panels' ones because it is a Static method, and I cannot iterate over the entryList.
 This modification required to modify all the extend methods to include a partitionID, and the Entry class to include a partitionID.  The Evaluation class has a public static variable DEFAULT_EVALUATION_ID=0 to cover the case of the sequential classifiers.

Edo